# openapi.yaml
openapi: 3.0.3

info:
  title: Customer Care API
  description: |
    REST API for the Customer Care domain.

    **Iteration 1 (current)** — One-time payment via Redis-backed account store.

    **Iteration 2 (planned)** — Oracle-backed persistence with full payment history and audit trail.

    > **Contract-first:** this file is the single source of truth.
    > Java request/response classes and API interfaces are **generated** from this spec
    > by the `openapi-generator-maven-plugin` during `mvn generate-sources`.
    > Do **not** edit the generated classes by hand.
  version: 1.0.0-SNAPSHOT
  contact:
    name: Customer Care Engineering

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Health
    description: Liveness / smoke-test endpoints
  - name: Payment
    description: One-time payment operations

paths:

  /hello:
    get:
      tags:
        - Health
      summary: Hello World
      description: >
        Simple smoke-test endpoint. Returns a greeting string to confirm the
        service is running and reachable.
      operationId: hello
      responses:
        '200':
          description: Service is up and running
          content:
            text/plain:
              schema:
                type: string
                example: Hello from customer-care-api!
        '500':
          $ref: '#/components/responses/InternalServerError'

  /one-time-payment:
    post:
      tags:
        - Payment
      summary: Process a one-time payment
      description: |
        Deducts the submitted payment amount **and** a tier-based match from the account balance.

        **Match tiers:**

        | Payment Amount   | Match % |
        |------------------|---------|
        | $0.01 – $9.99    | 1 %     |
        | $10.00 – $49.99  | 3 %     |
        | $50.00 and above | 5 %     |

        The next payment due date is set to **15 calendar days** from today.
        If that date falls on a **Saturday or Sunday** it is moved forward to the
        following **Monday**.
      operationId: oneTimePayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OneTimePaymentRequest'
            examples:
              low-tier:
                summary: Low-tier match — 1 % ($5 payment)
                value:
                  userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  paymentAmount: 5.00
              mid-tier:
                summary: Mid-tier match — 3 % ($10 payment)
                value:
                  userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  paymentAmount: 10.00
              high-tier:
                summary: High-tier match — 5 % ($75 payment)
                value:
                  userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  paymentAmount: 75.00
      responses:
        '200':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OneTimePaymentResponse'
              examples:
                mid-tier:
                  summary: Mid-tier result — $10 payment, 3 % match, weekday due date
                  value:
                    userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    previousBalance: 100.00
                    paymentAmount: 10.00
                    matchPercentage: 3
                    matchAmount: 0.30
                    newBalance: 89.70
                    nextPaymentDueDate: "2022-03-29"
                high-tier:
                  summary: High-tier result — $75 payment, 5 % match, weekend-shifted due date
                  value:
                    userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    previousBalance: 500.00
                    paymentAmount: 75.00
                    matchPercentage: 5
                    matchAmount: 3.75
                    newBalance: 421.25
                    nextPaymentDueDate: "2022-04-25"
        '400':
          description: Validation failure or invalid payment amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-04-13T10:15:30Z"
                status: 400
                error: "Bad Request"
                message: "Validation failed"
                errors:
                  - "paymentAmount: paymentAmount must be greater than 0"
        '404':
          description: Account not found for the supplied userId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-04-13T10:15:30Z"
                status: 404
                error: "Not Found"
                message: "Account not found for userId: 3fa85f64-5717-4562-b3fc-2c963f66afa6"
        '500':
          $ref: '#/components/responses/InternalServerError'

components:

  responses:
    InternalServerError:
      description: Unexpected server-side error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2026-04-13T10:15:30Z"
            status: 500
            error: "Internal Server Error"
            message: "An unexpected error occurred"

  schemas:

    OneTimePaymentRequest:
      type: object
      description: Request body for `POST /one-time-payment`.
      required:
        - userId
        - paymentAmount
      properties:
        userId:
          type: string
          minLength: 1
          description: Identifies the account to apply the payment to.
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        paymentAmount:
          type: number
          minimum: 0.01
          description: Payment amount in USD; must be ≥ $0.01.
          example: 10.00

    OneTimePaymentResponse:
      type: object
      description: Result of a successful one-time payment.
      required:
        - userId
        - previousBalance
        - paymentAmount
        - matchPercentage
        - matchAmount
        - newBalance
        - nextPaymentDueDate
      properties:
        userId:
          type: string
          description: Echo of the requesting user.
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        previousBalance:
          type: number
          description: Balance before this payment.
          example: 100.00
        paymentAmount:
          type: number
          description: The submitted payment amount.
          example: 10.00
        matchPercentage:
          type: integer
          description: Applied match tier — 1, 3, or 5.
          example: 3
        matchAmount:
          type: number
          description: Dollar value of the match deducted from the balance.
          example: 0.30
        newBalance:
          type: number
          description: Balance after payment + match deduction.
          example: 89.70
        nextPaymentDueDate:
          type: string
          format: date
          description: Weekend-adjusted due date (ISO-8601).
          example: "2022-03-29"

    ErrorResponse:
      type: object
      description: Standard error envelope returned for all non-2xx responses.
      required:
        - timestamp
        - status
        - error
        - message
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred (ISO-8601).
          example: "2026-04-13T10:15:30Z"
        status:
          type: integer
          description: HTTP status code.
          example: 500
        error:
          type: string
          description: HTTP status reason phrase.
          example: Internal Server Error
        message:
          type: string
          description: Human-readable description of the error.
          example: An unexpected error occurred
        errors:
          type: array
          description: Field-level validation messages — populated for 400 validation failures only.
          items:
            type: string
          example:
            - "paymentAmount: paymentAmount must be greater than 0"